refactor: Ladder와 LadderMaker 통합
