+++ src/main/java/ladder/domain/ladder/Ladder.java
@@ -1,13 +1,36 @@
 package ladder.domain.ladder;
 
+import ladder.domain.ladder.builder.BlockGenerator;
+import ladder.domain.ladder.exception.LadderLengthException;
+
+import java.util.ArrayList;
 import java.util.List;
 
 public class Ladder {
 
     private final List<Line> lines;
+    private final BlockGenerator blockGenerator;
 
-    public Ladder(final List<Line> lines) {
-        this.lines = List.copyOf(lines);
+    public Ladder(final BlockGenerator blockGenerator, final int playerNumber, final int height) {
+        validateLadderLength(playerNumber, height);
+        this.blockGenerator = blockGenerator;
+        this.lines = makeLines(playerNumber, height);
+    }
+
+    private void validateLadderLength(final int playerNumber, final int height) {
+        if (height < playerNumber - 1) {
+            throw new LadderLengthException();
+        }
+    }
+
+    public List<Line> makeLines(final int playerNumber, final int height) {
+        List<Line> lines = new ArrayList<>();
+        Line line = new Line(playerNumber, blockGenerator);
+
+        for (int i = 0; i < height; i++) {
+            lines.add(line);
+        }
+        return lines;
     }
 
     public List<Line> getLines() {
